Use gen_statem for party servers
Because when restoring from state means that while the system is
starting up, the party is in an incomplete state and should behave
differently.
gen_statem also offer interesting features like event postponing so as
to avoid dropping stuff while partially initialised, or state timeout to
abort restoration if something goes wrong (or not quick enough).